refactor: input 폴더구조 수정 및 NicknameInput label 연결
