+++ src/components/common/input/ProductNameInput.tsx
@@ -1,4 +1,4 @@
-import { StyledProductNameInput } from "./StyledProductnameInput";
+import { StyledProductNameInput } from "./Styled/StyledProductnameInput";
 
 export default function ProductNameInput() {
   return <StyledProductNameInput placeholder="상품명" />;
