+++ src/components/common/input/Styled/StyledNicknameInput.tsx
