+++ src/components/common/input/Styled/StyledTextBox.tsx
